pack individual funcs into one convenant
